requirements.txt
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
google-auth==2.40.3
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.179.0
google-cloud-aiplatform==1.78.0
vertexai==1.78.0
google-generativeai==0.8.5